Add cast to prevent warnings on 32 bit devices

+++ WTADataExample/WTADataExampleTests/WTADataExampleTests.m
@@ -222,7 +222,7 @@
         count += [items count];
     }
     
-    NSLog(@"Starting number of objects: %lu", count);
+    NSLog(@"Starting number of objects: %lu", (unsigned long)count);
  
     XCTestExpectation *expectation = [self expectationWithDescription:@"testDeleteAllDataWithCompletion"];
